EHTImages

Documentation for EHTImages.

EHTImages.ehtimage_metadata_compat

EHTImages.ehtimage_metadata_default

EHTImages.ehtimage_metadata_type

EHTImages.ncd_image_defaultgroup

EHTImages.ncd_image_dimnames

EHTImages.ncd_image_metadata_typeconv

EHTImages.ncd_image_vardims

EHTImages.ncd_image_varnames

EHTImages.ncd_image_vartypes

EHTImages.DDImage

EHTImages.DataStorageType

EHTImages.IsDiskData

EHTImages.NotDiskData

Base.close

Base.isopen

Base.iswritable

EHTImages.close!

EHTImages.concat

EHTImages.convolve

EHTImages.convolve!

EHTImages.create_ddimage

EHTImages.create_ddimage

EHTImages.create_ncimage

EHTImages.default_metadata

EHTImages.define_ncd_image_dimensions!

EHTImages.define_ncd_image_variables!

EHTImages.get_bconv

EHTImages.get_uvgrid

EHTImages.get_uvgrid

EHTImages.get_xygrid

EHTImages.get_xygrid

EHTImages.hdulist2ant

EHTImages.hdulist2freq

EHTImages.hdulist2hdus

EHTImages.hdulist2metadata

EHTImages.isdiskdata

EHTImages.load

EHTImages.load_fits

EHTImages.load_fits

EHTImages.load_fits

EHTImages.load_image

EHTImages.open!

EHTImages.save_fits!

EHTImages.save_netcdf

EHTImages.save_netcdf!

EHTImages.set_ncd_image_metadata!

EHTImages.ehtimage_metadata_compat — Constant

ehtimage_metadata_compat::NamedTuple

A tuple of available values for some of keys in ehtimage_metadata_default.

EHTImages.ehtimage_metadata_default — Constant

ehtimage_metadata_default::NamedTuple

A tuple for the default metadata keys and values for AbstractEHTImage.

EHTImages.ehtimage_metadata_type — Constant

ehtimage_metadata_default::NamedTuple

A tuple of types for metadata keys in ehtimage_metadata_default.

EHTImages.ncd_image_defaultgroup — Constant

ncd_image_defaultgroup::String

Default group name of the EHT Image NetCDF4 format.

EHTImages.ncd_image_dimnames — Constant

ncd_image_dimnames::NamedTuple

A named tuple relating Symbols to actual strings for the dimension of the EHT Image NetCDF4 format. keys are :x, :y for x and y axis, :p for polarization, :f for frequency, :t for time.

EHTImages.ncd_image_metadata_typeconv — Constant

ncd_image_metadata_typeconv::NamedTuple

A named tuple relating Symbols to types of the corresponding variables of the EHT Image NetCDF4 format.

EHTImages.ncd_image_vardims — Constant

ncd_image_vartypes::NamedTuple

A named tuple relating Symbols to types of the corresponding variables of the EHT Image NetCDF4 format.

EHTImages.ncd_image_varnames — Constant

ncd_image_varnames::NamedTuple

A named tuple relating Symbols to names of the corresponding variables of the EHT Image NetCDF4 format.

EHTImages.ncd_image_vartypes — Constant

ncd_image_vartypes::NamedTuple

A named tuple relating Symbols to types of the corresponding variables of the EHT Image NetCDF4 format.

EHTImages.DDImage — Type

struct DDImage <: AbstractEHTImage

<: AbstractEHTImage

A data type for five dimensional images of which data are all stored in the memory. This format relies on DimensionalData.DimArray to provide an easy access of data through many useful methods in Dimensional Data. Note that this data type is immutable.

Fields

data::Array{Float64, 5}: The five dimensional intensity disbrituions. Alias to dimarray.data.

pol::Vector{String}: The polarization code, giving the polarization axis (:p) of dimarray. Alias to dimarray.dims[3].val.data.

freq::Vector{Float64}: The central frequency in Hz, giving the frequency axis (:f) of dimarray. Alias to dimarray.dims[4].val.data.

mjd::Vector{Float64}: The modified Julian date, giving the time axis (:t) of dimarray. Alias to dimarray.dims[5].val.data.

metadata::OrderedDict: metadata.

dimarray::DimArray{Float64, 5}. DimArray stroing all of data, pol, freq, mjd and metadata.

EHTImages.DataStorageType — Type

DataStorageType

Internal type for specifying the nature of the location of data.

EHTImages.IsDiskData — Type

struct IsDiskData <: EHTImages.DataStorageType

Defines a trait that a states that data is disk based.

EHTImages.NotDiskData — Type

struct NotDiskData <: EHTImages.DataStorageType

Defines a trait that a states that data is memory based.

Base.close — Method

close(image::NCImage)

Close the access to the associated NetCDF4 file. This function is an alias to close!(image).

Base.isopen — Method

isopen(image::AbstractEHTImage)

Check if data is accessible, return true for accessible data and false if data is not accessible. This is relevant if image is based on disk data.

Base.iswritable — Method

isopen(image::AbstractEHTImage)

Check if data is accessible, return true for accessible data and false if data is not accessible. This is relevant if image is based on disk data.

EHTImages.close! — Method

close!(image::NCImage)

Close the access to the associated NetCDF4 file.

EHTImages.concat — Method

concat(dataset1::DimStack, dataset2::DimStack)

EHTImages.convolve! — Method

convolve!(image, model; ex)

EHTImages.convolve — Method

convolve(image, model; ex)

-> AbstractEHTImage

Convolve the input image with a given model, and return the convolved image.

Arguments

image::AbstractEHTImage: The input image. It must be not disk-based.

model::EHTModels.AbstractModel: The model to be used as the convolution kernel.

ex=SequentialEx() An executor of FLoops.jl.

EHTImages.create_ddimage — Method

create_ddimage(dimarray::DimArray)

Create DDImage instance from a given DimArray.

EHTImages.create_ddimage — Method

create_image(nx, dx, angunit; keywords) -> DDImage
create_ddimage(nx, dx, angunit; keywords) -> DDImage

Create and return a blank DDImage object.

Arguments

nx::Integer: the number of pixels along with the horizontal axis. Must be positive.

dx::Real: the pixel size of the horizontal axis. Must be positive.

angunit::Union{Unitful.Quantity, Unitful.Units or String}=rad: the angular unit for dx and dy.

Keywords

ny::Real=nx: the number of pixels along with the vertical axis. Must be positive.

dy::Real=dx: the pixel size of the vertical axis. Must be positive.

ixref::Real=(nx + 1) / 2, iyref::Real=(ny + 1) / 2: index of the reference pixels along with the horizontal and vertical axises, respectively. Default values set to the center of the field of the view.

pol::Symbol=:single: number of polarizations. Availables are :single or :full (i.e. four) polarizations.

freq::Vector{Float64}=[1.0]: a vector for frequencies in the unit of Hz

mjd::Vector{Float64}=[0.0]: a vector for time in the unit of MJD.

metadata::AbstractDict=default_metadata(NCImage()): other metadata. Note that the above keywords and arguments will overwrite the values of the conflicting keys in this metadata argument.

EHTImages.create_ncimage — Method

create_ncimage(filename, nx, dx, angunit; keywords) -> NCImage

Create a blank NCImage object. Return NCImage data loaded with :read mode.

Arguments

filename::AbstractString: NetCDF4 file where image data will be created.

nx::Integer: the number of pixels along with the horizontal axis. Must be positive.

dx::Real: the pixel size of the horizontal axis. Must be positive.

angunit::Union{Unitful.Quantity, Unitful.Units or String}=rad: the angular unit for dx and dy.

Keywords

ny::Real=nx: the number of pixels along with the vertical axis. Must be positive.

dy::Real=dx: the pixel size of the vertical axis. Must be positive.

ixref::Real=(nx + 1) / 2, iyref::Real=(ny + 1) / 2: index of the reference pixels along with the horizontal and vertical axises, respectively. Default values set to the center of the field of the view.

pol::Symbol=:single: number of polarizations. Availables are :single or :full (i.e. four) polarizations.

freq::Vector{Float64}=[1.0]: a vector for frequencies in the unit of Hz

mjd::Vector{Float64}=[0.0]: a vector for time in the unit of MJD.

metadata::AbstractDict=default_metadata(NCImage()): other metadata. Note that the above keywords and arguments will overwrite the values of the conflicting keys in this metadata argument.

mode::Symbol=:create: The access mode to NCDataset. Available modes are :read, :append, :create. See help for EHTNCDBase.ncdmodes for details.

group::AbstractString=EHTImage.ncd_image_defaultgroup: The group of the image data in the input NetCDF4 file.

EHTImages.default_metadata — Method

default_metadata(dataset) -> OrderedDict

Return the default metadata of the given dataset.

EHTImages.define_ncd_image_dimensions! — Function

define_ncd_image_dimensions!(ncd[, nx, ny, np, nf, nt])

Define NetCDF4 dimensions based on the given size of the image data.

EHTImages.define_ncd_image_variables! — Method

define_ncd_image_variables!(ncd)

Define NetCDF4 variables based on EHT NetCDF4 Image Format.

EHTImages.get_bconv — Method

get_bconv

get a conversion factor from Jy/pixel (used in AbstractEHTImage.data) to an arbitrary unit for the intensity. fluxunit is for the unit of the flux density (e.g. Jy, mJy, μJy) or brightness temperture (e.g. K), while saunit is for the unit of the solid angle (pixel, beam, mas, μJy).

EHTImages.get_uvgrid — Function

get_uvgrid(metadata, dofftshift=true)

returning u and v grids corresponding to the image field of view and pixel size.

EHTImages.get_uvgrid — Function

get_uvgrid(image, dofftshift=true)

returning u and v grids corresponding to the image field of view and pixel size.

EHTImages.get_xygrid — Function

get_xygrid --> Tuple{StepRangeLen, StepRangeLen}

Returning 1-dimensional StepRangeLen objects for the grids along with x and y axis in the given angular unit specified by angunit.

Arguments

metadata::Dict{Symbol, Any}-like: Input dictionary.

angunit::Union{Unitful.Quantity,Unitful.Units,String}=rad: Angular units of the output pixel grids.

EHTImages.get_xygrid — Function

get_xygrid

Returning 1-dimensional StepRange objects for the grids along with x and y axis in the given angular unit specified by angunit.

EHTImages.hdulist2ant — Method

hdulist2ant(antab)

EHTImages.hdulist2freq — Method

hdulist2freq(ghdu, antab, fqtab)

EHTImages.hdulist2hdus — Method

hdulist2hdus(hdulist)

Read the given hdulist (output of pyfits.io.open) and return GroupHDU and HDUs of AIPS AN/FQ Tables.

EHTImages.hdulist2metadata — Method

hdulist2metadata(ghdu, antab)

EHTImages.isdiskdata — Method

isdiskdata(data)

Determines whether the data is disk-based or memory-based. Return IsDiskData() if data is disk-based, while return NotDiskData() if data is memory-based.

EHTImages.load — Method

load(image::NCImage) --> DDImage

Load image data from the input disk image to memory.

Arguments

image::NCImage: Input NCImage.

EHTImages.load_fits — Function

load_fits(filename::AbstractString, hduid::Integer=1)

Load the input FITS image into DDImage (in-memory image data).

Arguments

filename::AbstractString: name of the input FITS file

hduid::Integer=1: ID of the HDU to be loaded. Default to the primary HDU.

EHTImages.load_fits — Function

load_fits(filename::FITS, hduid::Integer=1)

Load the input FITS image into DDImage (in-memory image data).

Arguments

fits::FITS: the input FITS data

hduid::Integer=1: ID of the HDU to be loaded. Default to the primary HDU.

EHTImages.load_fits — Method

load_fits(hdu)

Load the input FITS image into DDImage (in-memory image data).

Arguments

hdu::ImageHDU: the input image HDU data

EHTImages.load_image — Method

load_image(filename; [group, mode]) -> NCImage

Load image data from the specified group in the given NetCDF4 file with the specified access mode.

Arguments

filename::AbstractString: The input NetCDF4 file.

group::AbstractString=EHTImage.ncd_image_defaultgroup The group of the image data in the input NetCDF4 file.

mode::Symbol=:read: The access mode to NCDataset. Available modes are :read, :append, :create. See help for EHTImage.ncdmodes for details.

EHTImages.open! — Function

open!(image[, mode])

Load image data from NCDataset specified in the input image object with the given access mode. If image data are already opened, it will close it and reload data again.

Arguments

image::NCImage: The input image object.

mode::Symbol=:read: The access mode to NCDataset. Available modes are :read, :append, :create. See help for EHTImage.ncdmodes for details.

EHTImages.save_fits! — Function

save_fits[!](image::AbstractEHTImage, filename::AbstractString, idx=(1, 1); fitstype::Symbol=:casa)

Saving the image into a FITS file in a specifed format.

Arguments

image::AbstractEHTImage: the input image

filename::AbstractString: the name of the output FITS file

idx: the index of the saved image. Should be (frequency index, time index). Default to (1,1).

Keywords

fitstype::Symbol: the format type of the output FITS. Availables are :casa (CASA compatible).

EHTImages.save_netcdf! — Method

save_netcdf!(image, filename; [mode, group])

Save image data to NetCDF4 format.

Arguments

image::AbstractEHTImage Input image data

filename::AbstractString: NetCDF4 file where image data will be saved.

mode::Symbol=:create: The access mode to NCDataset. Available modes are :read, :append, :create. See help for EHTNCDBase.ncdmodes for details.

group::AbstractString=EHTImage.ncd_image_defaultgroup: The group of the image data in the input NetCDF4 file.

EHTImages.save_netcdf — Method

save_netcdf(image, filename; [mode=:create, group="image"]) => NCImage

Save image data to NetCDF4 format. Saved data will be loaded and returned with :read access mode.

Arguments

image::AbstractEHTImage Input image data

filename::AbstractString: NetCDF4 file where image data will be saved.

mode::Symbol=:create: The access mode to NCDataset. Available modes are :read, :append, :create. See help for EHTNCDBase.ncdmodes for details.

group::AbstractString=EHTImage.ncd_image_defaultgroup: The group of the image data in the input NetCDF4 file.

EHTImages.set_ncd_image_metadata! — Method

set_ncd_image_metadata!(ncd)

Set NetCDF4 metadata based on EHT NetCDF4 Image Format.