EHTImages.ehtimage_metadata_compat

EHTImages.ehtimage_metadata_default

EHTImages.ehtimage_metadata_type

EHTImages.ncd_intensity_defaultgroup

EHTImages.ncd_intensity_dimnames

EHTImages.ncd_intensity_metadata_typeconv

EHTImages.ncd_intensity_vardims

EHTImages.ncd_intensity_varnames

EHTImages.ncd_intensity_vartypes

EHTImages.AbstractImageDataSet

EHTImages.AbstractIntensityImage

EHTImages.DataStorageType

EHTImages.DiskIntensityImage

EHTImages.IntensityImage

EHTImages.IsDiskData

EHTImages.NotDiskData

Base.close

Base.isopen

Base.iswritable

EHTImages.close!

EHTImages.convolve

EHTImages.convolve!

EHTImages.convolve_base!

EHTImages.copy_metadata!

EHTImages.default_metadata

EHTImages.default_metadata

EHTImages.define_ncd_intensity_dimensions!

EHTImages.define_ncd_intensity_variables!

EHTImages.diskintensityimage

EHTImages.get_bconv

EHTImages.get_uvgrid

EHTImages.get_uvgrid

EHTImages.get_xygrid

EHTImages.get_xygrid

EHTImages.intensityimage

EHTImages.isdiskdata

EHTImages.load

EHTImages.load_fits

EHTImages.load_image

EHTImages.open!

EHTImages.save_fits!

EHTImages.save_netcdf

EHTImages.save_netcdf!

EHTImages.set_ncd_intensity_metadata!

EHTImages.ehtimage_metadata_compat — Constant

ehtimage_metadata_compat::NamedTuple

A tuple of available values for some of keys in ehtimage_metadata_default.

EHTImages.ehtimage_metadata_default — Constant

ehtimage_metadata_default::NamedTuple

A tuple for the default metadata keys and values for AbstractIntensityImage.

EHTImages.ehtimage_metadata_type — Constant

ehtimage_metadata_default::NamedTuple

A tuple of types for metadata keys in ehtimage_metadata_default.

EHTImages.ncd_intensity_defaultgroup — Constant

ncd_intensity_defaultgroup::String

Default group name of the EHT Intensity Image NetCDF4 format.

EHTImages.ncd_intensity_dimnames — Constant

ncd_intensity_dimnames::NamedTuple

A named tuple relating Symbols to actual strings for the dimension of the EHT Intensity Image NetCDF4 format. keys are :x, :y for x and y axis, :p for polarization, :f for frequency, :t for time.

EHTImages.ncd_intensity_metadata_typeconv — Constant

ncd_intensity_metadata_typeconv::NamedTuple

A named tuple relating Symbols to types of the corresponding variables of the EHT Intensity Image NetCDF4 format.

EHTImages.ncd_intensity_vardims — Constant

ncd_intensity_vartypes::NamedTuple

A named tuple relating Symbols to types of the corresponding variables of the EHT Intensity Image NetCDF4 format.

EHTImages.ncd_intensity_varnames — Constant

ncd_intensity_varnames::NamedTuple

A named tuple relating Symbols to names of the corresponding variables of the EHT Intensity Image NetCDF4 format.

EHTImages.ncd_intensity_vartypes — Constant

ncd_intensity_vartypes::NamedTuple

A named tuple relating Symbols to types of the corresponding variables of the EHT Intensity Image NetCDF4 format.

EHTImages.AbstractImageDataSet — Type

abstract type AbstractImageDataSet

The abstract type for image data sets handled in this module. AbstractImageDataSet works as an Abstract Array-ish. To make it, each image type needs to have four following methods. (see: Julia Documentation for "Interfaces")

Mandatory Methods that need to be defined

default_metadata: Return the default metadata for the image data set.

isdiskdata: Determines whether the data is disk-based or memory-based. Return IsDiskData() if data is disk-based, while return NotDiskData() if data is memory-based.

isopen: Check if data is accessible, return true for accessible data and false if data is not accessible. This is relevant if image is based on disk data.

iswritable: Check if data is accessible, return true for accessible data and false if data is not accessible. This is relevant if image is based on disk data.

Methods provided

size: returning a tuple containing the dimension of AbstractImageDataSet.data

getindex: scalar or vector indexing

setindex!: scalar or vector indexing assignment

firstindex: returning the first index, used in X[begin]

lastindex: returning the last index, used in X[end]

IndexStyle: returning the index style

EHTImages.AbstractIntensityImage — Type

abstract type AbstractIntensityImage <: AbstractImageDataSet

This defines a basic interface for intensity images. It is a subtype of AbstractImageDataSet.

Mandatory fields

data: 5 dimensional array for intensity [x, y, polarization, frequency, time]

p: 1 dimensional array for polarization codes in string (coordinate for polarization axis)

f: 1 dimensional array for frequency in Hz (coordinate for frequency axis)

t: 1 dimensional array for time in modified Julian dates (coordinate for time axis)

metadata: Dict-like object to stock metadata

Mandatory methods need to be defined.

See also documentations for AbstractImageDataSet.

EHTImages.DataStorageType — Type

abstract type DataStorageType

Internal type for specifying the nature of the location of data.

EHTImages.DiskIntensityImage — Type

mutable struct DiskIntensityImage <: AbstractIntensityImage

A data type for five dimensional images of which data are all stored in disk using the NetCDF4 format. This format relies on NCDatasets to provide an easy access of data through many useful methods in the NCDatasets package. Note that this data type could be either mutable or immutable depending on the access mode to the NetCDF4 file.

filename::Any: name of the corresponding NetCDF4 file Default: nothing

group::Any: group name of the corresponding image data set Default: nothing

data::Any: five dimensional intensity disbrituion. Default: nothing

metadata::Any: metadata. Default: nothing

t::Any: polarization code, giving the parization axis (:p). Default: nothing

f::Any: central frequency in Hz, giving the frequency axis (:f). Default: nothing

p::Any: central modified Julian date, giving the time axis (:t). Default: nothing

dataset::Any: Default: nothing

EHTImages.IntensityImage — Type

struct IntensityImage <: AbstractIntensityImage

A data type for five dimensional images of which data are all stored in the memory. This format relies on EHTDimensionalData.DimArray to provide an easy access of data through many useful methods in EHTDimensionalData and its origin DimensionalData packages. Note that this data type is immutable.

dimstack::DimensionalData.DimStack: DimArray storing all of data, p, f, t and metadata.

data::Array{Float64, 5}: the five dimensional intensity disbrituion. Alias to dimstack.intensity.data.

metadata::OrderedCollections.OrderedDict{Symbol, Any}: metadata. Aliast to dimarray.metadata.

p::Vector{String}: the polarization code, giving the polarization axis (:p) of dimarray. Alias to dimarray.dims[3].val.data.

f::Vector{Float64}: the central frequency in Hz, giving the frequency axis (:f) of dimarray. Alias to dimarray.dims[4].val.data.

t::Vector{Float64}: the central modified Julian date, giving the time axis (:t) of dimarray. Alias to dimarray.dims[5].val.data.

EHTImages.IsDiskData — Type

struct IsDiskData <: EHTImages.DataStorageType

Defines a trait that a states that data is disk based.

EHTImages.NotDiskData — Type

struct NotDiskData <: EHTImages.DataStorageType

Defines a trait that a states that data is memory based.

Base.close — Method

close(image::DiskIntensityImage)

Close the access to the associated NetCDF4 file. This function is an alias to close!(image).

Base.isopen — Method

isopen(_::AbstractImageDataSet) -> Bool

Check if data is accessible, return true for accessible data and false if data is not accessible. This is relevant if image is based on disk data.

Base.iswritable — Method

iswritable(_::AbstractImageDataSet) -> Bool

Check if data is accessible, return true for accessible data and false if data is not accessible. This is relevant if image is based on disk data.

EHTImages.close! — Method

close!(image::DiskIntensityImage)

Close the access to the associated NetCDF4 file.

EHTImages.convolve! — Method

convolve!(
    image::AbstractIntensityImage,
    model::EHTModels.AbstractModel;
    ex
)

Convolve the input image with a given model.

Arguments

image::AbstractIntensityImage: The input image. It must be not disk-based.

model::EHTModels.AbstractModel: The model to be used as the convolution kernel.

ex=SequentialEx() An executor of FLoops.jl.

EHTImages.convolve — Method

convolve(
    image::AbstractIntensityImage,
    model::EHTModels.AbstractModel;
    ex
) -> AbstractIntensityImage

-> AbstractIntensityImage

Convolve the input image with a given model, and return the convolved image.

Arguments

image::AbstractIntensityImage: The input image. It must be not disk-based.

model::EHTModels.AbstractModel: The model to be used as the convolution kernel.

ex=SequentialEx() An executor of FLoops.jl.

EHTImages.convolve_base! — Method

convolve_base!(
    image::AbstractIntensityImage,
    model::EHTModels.AbstractModel;
    ex
) -> Any

Base function for convolving the input image with a given model.

Arguments

image::AbstractIntensityImage: The input image. It must be not disk-based.

model::EHTModels.AbstractModel: The model to be used as the convolution kernel.

ex=SequentialEx() An executor of FLoops.jl.

EHTImages.copy_metadata! — Method

copy_metadata!(image::AbstractIntensityImage, uvdataset::AbstractUVDataSet)

copy metadata from the given uvdataset.

EHTImages.default_metadata — Method

default_metadata(::Type{<:AbstractImageDataSet}) -> OrderedDict{Symbol, Any}
default_metadata(::AbstractImageDataSet) -> OrderedDict{Symbol, Any}

Return default metadata for the image data set.

EHTImages.default_metadata — Method

default_metadata(dataset) -> OrderedDict

Return the default metadata of the given dataset.

EHTImages.define_ncd_intensity_dimensions! — Function

define_ncd_intensity_dimensions!(ncd[, nx, ny, np, nf, nt])

Define NetCDF4 dimensions based on the given size of the image data.

EHTImages.define_ncd_intensity_variables! — Method

define_ncd_intensity_variables!(ncd)

Define NetCDF4 variables based on EHT NetCDF4 Image Format.

EHTImages.diskintensityimage — Method

diskintensityimage(filename, nx, dx, angunit; keywords) -> DiskIntensityImage

Create a blank DiskIntensityImage object. Return DiskIntensityImage data loaded with :read mode.

Arguments

filename::AbstractString: NetCDF4 file where image data will be created.

nx::Integer: the number of pixels along with the horizontal axis. Must be positive.

dx::Real: the pixel size of the horizontal axis. Must be positive.

angunit::Union{Unitful.Quantity, Unitful.Units or String}=rad: the angular unit for dx and dy.

Keywords

ny::Real=nx: the number of pixels along with the vertical axis. Must be positive.

dy::Real=dx: the pixel size of the vertical axis. Must be positive.

ixref::Real=(nx + 1) / 2, iyref::Real=(ny + 1) / 2: index of the reference pixels along with the horizontal and vertical axises, respectively. Default values set to the center of the field of the view.

pol::Symbol=:single: number of polarizations. Availables are :single or :full (i.e. four) polarizations.

freq::Vector{Float64}=[1.0]: a vector for frequencies in the unit of Hz

mjd::Vector{Float64}=[0.0]: a vector for time in the unit of MJD.

metadata::AbstractDict=default_metadata(AbstractIntensityImage): other metadata. Note that the above keywords and arguments will overwrite the values of the conflicting keys in this metadata argument.

mode::Symbol=:create: The access mode to NCDataset. Available modes are :read, :append, :create. See help for EHTNCDBase.ncdmodes for details.

group::AbstractString=EHTImage.ncd_intensity_defaultgroup: The group of the image data in the input NetCDF4 file.

EHTImages.get_bconv — Method

get_bconv

get a conversion factor from Jy/pixel (used in AbstractIntensityImage.data) to an arbitrary unit for the intensity. fluxunit is for the unit of the flux density (e.g. Jy, mJy, μJy) or brightness temperture (e.g. K), while saunit is for the unit of the solid angle (pixel, beam, mas, μJy).

EHTImages.get_uvgrid — Function

get_uvgrid(image, dofftshift=true)

returning u and v grids corresponding to the image field of view and pixel size.

EHTImages.get_uvgrid — Function

get_uvgrid(metadata, dofftshift=true)

returning u and v grids corresponding to the image field of view and pixel size.

EHTImages.get_xygrid — Function

get_xygrid --> Tuple{StepRangeLen, StepRangeLen}

Returning 1-dimensional StepRangeLen objects for the grids along with x and y axis in the given angular unit specified by angunit.

Arguments

metadata::Dict{Symbol, Any}-like: Input dictionary.

angunit::Union{Unitful.Quantity,Unitful.Units,String}=rad: Angular units of the output pixel grids.

EHTImages.get_xygrid — Function

get_xygrid

Returning 1-dimensional StepRange objects for the grids along with x and y axis in the given angular unit specified by angunit.

EHTImages.intensityimage — Method

intensityimage(nx, dx, angunit; keywords) -> IntensityImage

Create and return a blank IntensityImage object.

Arguments

nx::Integer: the number of pixels along with the horizontal axis. Must be positive.

dx::Real: the pixel size of the horizontal axis. Must be positive.

angunit::Union{Unitful.Quantity, Unitful.Units or String}=rad: the angular unit for dx and dy.

Keywords

ny::Real=nx: the number of pixels along with the vertical axis. Must be positive.

dy::Real=dx: the pixel size of the vertical axis. Must be positive.

ixref::Real=(nx + 1) / 2, iyref::Real=(ny + 1) / 2: index of the reference pixels along with the horizontal and vertical axises, respectively. Default values set to the center of the field of the view.

p::Symbol=:single: number of parizations. Availables are :single or :full (i.e. four) parizations.

f::Vector{Float64}=[1.0]: a vector for fuencies in the unit of Hz

t::Vector{Float64}=[0.0]: a vector for time in the unit of t.

metadata::AbstractDict=default_metadata(AbstractIntensityImage): other metadata. Note that the above keywords and arguments will overwrite the values of the conflicting keys in this metadata argument.

EHTImages.isdiskdata — Method

isdiskdata(_::AbstractImageDataSet) -> EHTImages.IsDiskData

Determines whether the data is disk-based or memory-based. Return IsDiskData() if data is disk-based, while return NotDiskData() if data is memory-based.

EHTImages.load — Method

load(image::NCImage) --> DDImage

Load image data from the input disk image to memory.

Arguments

image::NCImage: Input NCImage.

EHTImages.load_fits — Function

load_fits(filename::AbstractString, hduid::Integer=1) -> IntensityImage
load_fits(fits::FITS, hduid::Integer=1) -> IntensityImage
load_fits(hdu::ImageHDU) -> IntensityImage

Load the input FITS image into IntensityImage (in-memory image data).

Arguments

filename::AbstractString: name of the input FITS file

hduid::Integer=1: ID of the HDU to be loaded. Default to the primary HDU.

hdu::ImageHDU: HDU to be loaded.

EHTImages.load_image — Method

load_image(filename; [group, mode]) -> DiskIntensityImage

Load image data from the specified group in the given NetCDF4 file with the specified access mode.

Arguments

filename::AbstractString: The input NetCDF4 file.

group::AbstractString=EHTImage.ncd_intensity_defaultgroup The group of the image data in the input NetCDF4 file.

mode::Symbol=:read: The access mode to NCDataset. Available modes are :read, :append, :create. See help for EHTImage.ncdmodes for details.

EHTImages.open! — Function

open!(image[, mode])

Load image data from NCDataset specified in the input image object with the given access mode. If image data are already opened, it will close it and reload data again.

Arguments

image::DiskIntensityImage: The input image object.

mode::Symbol=:read: The access mode to NCDataset. Available modes are :read, :append, :create. See help for EHTImage.ncdmodes for details.

EHTImages.save_fits! — Function

save_fits[!](image::AbstractIntensityImage, filename::AbstractString, idx=(1, 1); fitstype::Symbol=:casa)

Saving the image into a FITS file in a specifed format.

Arguments

image::AbstractIntensityImage: the input image

filename::AbstractString: the name of the output FITS file

idx: the index of the saved image. Should be (frequency index, time index). Default to (1,1).

Keywords

fitstype::Symbol: the format type of the output FITS. Availables are :casa (CASA compatible).

EHTImages.save_netcdf! — Method

save_netcdf!(image, filename; [mode, group])

Save image data to NetCDF4 format.

Arguments

image::AbstractIntensityImage Input image data

filename::AbstractString: NetCDF4 file where image data will be saved.

mode::Symbol=:create: The access mode to NCDataset. Available modes are :read, :append, :create. See help for EHTNCDBase.ncdmodes for details.

group::AbstractString=EHTImage.ncd_intensity_defaultgroup: The group of the image data in the input NetCDF4 file.

EHTImages.save_netcdf — Method

save_netcdf(image, filename; [mode=:create, group="image"]) => DiskIntensityImage

Save image data to NetCDF4 format. Saved data will be loaded and returned with :read access mode.

Arguments

image::AbstractIntensityImage Input image data

filename::AbstractString: NetCDF4 file where image data will be saved.

mode::Symbol=:create: The access mode to NCDataset. Available modes are :read, :append, :create. See help for EHTNCDBase.ncdmodes for details.

group::AbstractString=EHTImage.ncd_intensity_defaultgroup: The group of the image data in the input NetCDF4 file.

EHTImages.set_ncd_intensity_metadata! — Method

set_ncd_intensity_metadata!(ncd)

Set NetCDF4 metadata based on EHT NetCDF4 Image Format.